Move classes that depend on DSv2 in Spark 2.x to spark2.
