+++ spark/src/main/java/org/apache/iceberg/spark/source/RowDataRewriter.java
@@ -23,22 +23,23 @@ import java.io.Serializable;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 import java.util.stream.Collectors;
 import org.apache.iceberg.CombinedScanTask;
 import org.apache.iceberg.DataFile;
 import org.apache.iceberg.FileFormat;
 import org.apache.iceberg.PartitionSpec;
-import org.apache.iceberg.SchemaParser;
+import org.apache.iceberg.Schema;
 import org.apache.iceberg.Table;
 import org.apache.iceberg.TableProperties;
 import org.apache.iceberg.encryption.EncryptionManager;
 import org.apache.iceberg.io.FileIO;
+import org.apache.iceberg.io.LocationProvider;
 import org.apache.iceberg.spark.SparkSchemaUtil;
 import org.apache.spark.TaskContext;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.broadcast.Broadcast;
 import org.apache.spark.sql.catalyst.InternalRow;
-import org.apache.spark.sql.sources.v2.writer.DataWriter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -48,58 +49,70 @@ public class RowDataRewriter implements Serializable {
 
   private static final Logger LOG = LoggerFactory.getLogger(RowDataRewriter.class);
 
-  private final Broadcast<FileIO> fileIO;
+  private final Schema schema;
+  private final PartitionSpec spec;
+  private final Map<String, String> properties;
+  private final FileFormat format;
+  private final Broadcast<FileIO> io;
   private final Broadcast<EncryptionManager> encryptionManager;
-  private final String tableSchema;
+  private final LocationProvider locations;
   private final String nameMapping;
-  private final Writer.WriterFactory writerFactory;
   private final boolean caseSensitive;
 
   public RowDataRewriter(Table table, PartitionSpec spec, boolean caseSensitive,
-                         Broadcast<FileIO> fileIO, Broadcast<EncryptionManager> encryptionManager,
-                         long targetDataFileSizeInBytes) {
-    this.fileIO = fileIO;
+                         Broadcast<FileIO> io, Broadcast<EncryptionManager> encryptionManager) {
+    this.schema = table.schema();
+    this.spec = spec;
+    this.locations = table.locationProvider();
+    this.properties = table.properties();
+    this.io = io;
     this.encryptionManager = encryptionManager;
 
     this.caseSensitive = caseSensitive;
-    this.tableSchema = SchemaParser.toJson(table.schema());
     this.nameMapping = table.properties().get(DEFAULT_NAME_MAPPING);
 
     String formatString = table.properties().getOrDefault(
         TableProperties.DEFAULT_FILE_FORMAT, TableProperties.DEFAULT_FILE_FORMAT_DEFAULT);
-    FileFormat fileFormat = FileFormat.valueOf(formatString.toUpperCase(Locale.ENGLISH));
-    this.writerFactory = new Writer.WriterFactory(spec, fileFormat, table.locationProvider(), table.properties(),
-        fileIO, encryptionManager, targetDataFileSizeInBytes, table.schema(), SparkSchemaUtil.convert(table.schema()));
+    this.format = FileFormat.valueOf(formatString.toUpperCase(Locale.ENGLISH));
   }
 
   public List<DataFile> rewriteDataForTasks(JavaRDD<CombinedScanTask> taskRDD) {
-    JavaRDD<Writer.TaskCommit> taskCommitRDD = taskRDD.map(this::rewriteDataForTask);
+    JavaRDD<TaskResult> taskCommitRDD = taskRDD.map(this::rewriteDataForTask);
 
     return taskCommitRDD.collect().stream()
         .flatMap(taskCommit -> Arrays.stream(taskCommit.files()))
         .collect(Collectors.toList());
   }
 
-  private Writer.TaskCommit rewriteDataForTask(CombinedScanTask task) throws Exception {
+  private TaskResult rewriteDataForTask(CombinedScanTask task) throws Exception {
     TaskContext context = TaskContext.get();
-
-    RowDataReader dataReader = new RowDataReader(task, SchemaParser.fromJson(tableSchema),
-        SchemaParser.fromJson(tableSchema), nameMapping, fileIO.value(),
-        encryptionManager.value(), caseSensitive);
-
     int partitionId = context.partitionId();
     long taskId = context.taskAttemptId();
-    DataWriter<InternalRow> dataWriter = writerFactory.createDataWriter(partitionId, taskId, 0);
+
+    RowDataReader dataReader = new RowDataReader(
+        task, schema, schema, nameMapping, io.value(), encryptionManager.value(), caseSensitive);
+
+    SparkAppenderFactory appenderFactory = new SparkAppenderFactory(
+        properties, schema, SparkSchemaUtil.convert(schema));
+    OutputFileFactory fileFactory = new OutputFileFactory(
+        spec, format, locations, io.value(), encryptionManager.value(), partitionId, taskId);
+
+    BaseWriter writer;
+    if (spec.fields().isEmpty()) {
+      writer = new UnpartitionedWriter(spec, format, appenderFactory, fileFactory, io.value(), Long.MAX_VALUE);
+    } else {
+      writer = new PartitionedWriter(spec, format, appenderFactory, fileFactory, io.value(), Long.MAX_VALUE, schema);
+    }
 
     try {
       while (dataReader.next()) {
         InternalRow row = dataReader.get();
-        dataWriter.write(row);
+        writer.write(row);
       }
 
       dataReader.close();
       dataReader = null;
-      return (Writer.TaskCommit) dataWriter.commit();
+      return writer.complete();
 
     } catch (Throwable originalThrowable) {
       try {
@@ -111,7 +124,7 @@ public class RowDataRewriter implements Serializable {
         if (dataReader != null) {
           dataReader.close();
         }
-        dataWriter.abort();
+        writer.abort();
         LOG.error("Aborted commit for partition {} (task {}, attempt {}, stage {}.{})",
             partitionId, taskId, context.taskAttemptId(), context.stageId(), context.stageAttemptNumber());
