+++ spark/src/main/java/org/apache/iceberg/spark/source/ThreeColumnRecord.java
