+++ spark2/src/test/java/org/apache/iceberg/spark/source/TestStructuredStreaming.java
